% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny_metagene.R
\name{shiny_metagene}
\alias{shiny_metagene}
\title{Launch interactive metagene session}
\usage{
shiny_metagene()
}
\description{
Launch interactive metagene session
}
\value{
Interactive metagene session
}
\examples{
if (interactive()) {
shiny_metagene()
}
}

